deployed face recognition with ovms
