+++ src/api/face_recognition_api.py
@@ -19,8 +19,7 @@ from PIL import Image
 
 from src.api.base_api import BaseAPI
 from src.engine.detector.ssd_detector import SSDFaceDetector
-# from src.engine.recognizer.facenet_recognizer import FaceNetRecognizer
-from src.engine.recognizer.facenet_recognizer2 import FaceNetRecognizer
+from src.engine.recognizer.facenet_recognizer import FaceNetRecognizer
 from src.schema.auth_schema import CurrentUser
 from src.schema.face_recognition_schema import FaceEmbeddingSchema
 from src.utils.exception import APIExceptions
@@ -52,7 +51,6 @@ class FaceRecognitionAPI(BaseAPI):
         self.face_detection.build_model()
 
         self.face_recognition = FaceNetRecognizer()
-        self.face_recognition.build_model()
 
     def setup(self) -> None:
         """Setup API Endpoints."""
@@ -161,26 +159,13 @@ class FaceRecognitionAPI(BaseAPI):
                 raise exceptions.BadRequest("Multiple faces detected")
 
             # recognize face
-            ground_truths = await self.mongodb.get_face_gts(user_id=user.id)
+            ground_truths = await self.mongodb.get_face_gts(user_id=str(user.id))
             preds = self.face_recognition.predict(
                 face=face_dets[0].face,
                 ground_truths=ground_truths,
                 dist_method=self.cfg.engine.recognizer.dist_method,
                 dist_threshold=self.cfg.engine.recognizer.min_dist_thres,
             )
-
-
-            # recognize face
-            # face_embd = self.face_recognition.get_embedding(face_dets[0].face)
-            # face_embd = FaceEmbeddingSchema(
-            #     user_id=user.id,
-            #     embedding=face_embd
-            #     )
-            # face_embd = await self.mongodb.find_face(
-            #     face_embd=face_embd,
-            #     min_dist_thres=self.cfg.engine.recognizer.min_dist_thres,
-            #     dist_method=self.cfg.engine.recognizer.dist_method,
-            # )
 
             end_request = t.now_iso(utc=True)
             log.log(
@@ -191,7 +176,7 @@ class FaceRecognitionAPI(BaseAPI):
             return {
                 "result": {
                     "status": "success",
-                    "name": face_embd.name,
+                    "name": preds.name,
                 }
             }
